+++ client/src/ts/fetchers/character.ts
@@ -1,6 +1,6 @@
 import { Fetcher } from '../utils/Fetcher'
 
-export async function fetchCharacters(pageSize: number = 100, pageNumber: number = 1) {
+export async function fetchCharacters(pageSize: number = 50, pageNumber: number = 1) {
     const urlRoot = `https://anapioficeandfire.com/api/`
     const characterUrlRoot = `${urlRoot}/characters`
     const charactersUrl = `${characterUrlRoot}/?pageSize=${pageSize}&page=${pageNumber}`